+++ doczrc.js
@@ -1,0 +1,9 @@
+export default {
+  typescript: true,
+  ignore: ["README.md", "CHANGELOG.md"],
+  base: '/weui-react/',
+  menu: [
+    'Getting Started',
+  ],
+  dest: './docs'
+}