+++ src/WeUIFormRadio/WeUIFormRadio.tsx
@@ -1,5 +1,9 @@
 import React, { FC, forwardRef } from "react";
-import { FormControlLabel, Radio, FormControlLabelProps } from "@material-ui/core";
+import {
+  FormControlLabel,
+  Radio,
+  FormControlLabelProps
+} from "@material-ui/core";
 
 interface Props extends Omit<FormControlLabelProps, "control"> {
   label: string;
@@ -12,9 +16,12 @@ const WeUIFormRadio: FC<Props> = forwardRef(
         {...props}
         ref={ref}
         value={value}
-        control={<Radio />}
+        control={
+          <Radio
+            style={{color: 'var(--weui-BRAND)'}}
+          />
+        }
         label={label}
-        style={{ ...props.style, marginLeft: 0 }}
       />
     );
   }